requirements.txt
anyio==3.6.1
black==22.8.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.82.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
